Extract `getSeriesEventView`
