+++ static/app/views/starfish/queries/getSeriesEventView.tsx
@@ -1,0 +1,45 @@
+import sortBy from 'lodash/sortBy';
+
+import type {PageFilters} from 'sentry/types';
+import {intervalToMilliseconds} from 'sentry/utils/dates';
+import EventView from 'sentry/utils/discover/eventView';
+import {parseFunction} from 'sentry/utils/discover/fields';
+import {DiscoverDatasets} from 'sentry/utils/discover/types';
+import type {MutableSearch} from 'sentry/utils/tokenizeSearch';
+import {getIntervalForMetricFunction} from 'sentry/views/performance/database/getIntervalForMetricFunction';
+import {DEFAULT_INTERVAL} from 'sentry/views/performance/database/settings';
+
+export function getSeriesEventView(
+  search: MutableSearch | undefined,
+  pageFilters: PageFilters,
+  yAxis: string[]
+) {
+  // Pick the highest possible interval for the given yAxis selection. Find the ideal interval for each function, then choose the largest one. This results in the lowest granularity, but best performance.
+  const interval = sortBy(
+    yAxis.map(yAxisFunctionName => {
+      const parseResult = parseFunction(yAxisFunctionName);
+
+      if (!parseResult) {
+        return DEFAULT_INTERVAL;
+      }
+
+      return getIntervalForMetricFunction(parseResult.name, pageFilters.datetime);
+    }),
+    result => {
+      return intervalToMilliseconds(result);
+    }
+  ).at(-1);
+
+  return EventView.fromNewQueryWithPageFilters(
+    {
+      name: '',
+      query: search?.formatString() ?? undefined,
+      fields: [],
+      yAxis,
+      dataset: DiscoverDatasets.SPANS_METRICS,
+      interval,
+      version: 2,
+    },
+    pageFilters
+  );
+}
